% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-change-password.R
\name{changePasswordUI}
\alias{changePasswordUI}
\title{Change-password UI}
\usage{
changePasswordUI(id)
}
\arguments{
\item{id}{The Shiny module ID.}
}
\value{
A `shiny::wellPanel()` to place in your UI.
}
\description{
Change-password UI
}
