% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-sessions.R
\name{revoke_session}
\alias{revoke_session}
\title{Revoke an SSO session}
\usage{
revoke_session(con, cookie_value)
}
\arguments{
\item{con}{A DBI connection or pool object.}

\item{cookie_value}{The cookie value identifying the session to revoke.}
}
\value{
Invisibly `TRUE`/`FALSE`.
}
\description{
Revoke an SSO session
}
