% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-sessions-s3.R
\name{revoke_session_s3}
\alias{revoke_session_s3}
\title{Revoke an SSO session (S3-backed)}
\usage{
revoke_session_s3(cookie_value)
}
\value{
Invisibly `TRUE`/`FALSE`.
}
\description{
Revoke an SSO session (S3-backed)
}
