% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-identity.R
\name{set_user_password}
\alias{set_user_password}
\title{Set a user's password}
\usage{
set_user_password(con, user_id, new_password)
}
\arguments{
\item{con}{A DBI connection or pool object.}

\item{user_id}{The user's ID.}

\item{new_password}{The new plaintext password to hash and store.}
}
\value{
The result of the underlying `DBI::dbExecute()` call.
}
\description{
Set a user's password
}
