% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-identity.R
\name{verify_user_password}
\alias{verify_user_password}
\title{Verify a user's current password}
\usage{
verify_user_password(con, user_id, password)
}
\arguments{
\item{con}{A DBI connection or pool object.}

\item{user_id}{The user's ID.}

\item{password}{The plaintext password to verify.}
}
\value{
`TRUE`/`FALSE`.
}
\description{
Verify a user's current password
}
